backend/sdmay20_42/requirements.txt
djoser
django
djangorestframework
djangorestframework_simplejwt
mysqlclient
